requirements.txt
mss==6.1.0
numpy==1.22.0
opencv-python==4.5.5.62
PyDirectInput==1.0.4
pynput==1.7.4
six==1.16.0
pyautogui==0.9.53
Pillow==9.0.0
pywinauto==0.6.6
tk==0.1.0
pywin32==303
opencv-contrib-python==4.5.5.62
matplotlib==3.5.1
scikit-image==0.19.1
tensorflow==1.8.0
